Refactor: 분리된 클래스 임포트 수정
