+++ src/main/java/racingcar/validation/ValidationName.java
@@ -3,16 +3,13 @@ package racingcar.validation;
 
 import java.util.ArrayList;
 
-public class Validation {
+public class ValidationName {
 
     public static void validateName(String carName) {
         String regex = "^[a-zA-Z0-9가-힣]+$";
-        try {
-            isNameEmpty(carName);
-            validateNameInput(carName, regex);
-            validateNameLength(carName);
-        } catch (IllegalArgumentException e) {}
-
+        isNameEmpty(carName);
+        validateNameInput(carName, regex);
+        validateNameLength(carName);
     }
 
     public static void isNameEmpty(String carName) {
@@ -36,23 +33,6 @@ public class Validation {
     public static void validateSameName(ArrayList<String> carNameArray, String carName) {
         if (carNameArray.contains(carName)) {
             throw new IllegalArgumentException("동일한 이름은 작성할 수 없습니다.");
-        }
-    }
-
-    public static int validateGameCount(String gameCount) {
-        int attempt;
-        try {
-            attempt = Integer.parseInt(gameCount);
-            validNegativeNumber(attempt);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("0~2^31-1의 정수로만 횟수를 입력하세요(문자, 소수 불가).");
-        }
-        return attempt;
-    }
-
-    public static void validNegativeNumber(int attempt) {
-        if (attempt < 0) {
-            throw new IllegalArgumentException("음수로는 시도 횟수를 측정하지 못합니다.");
         }
     }
 }
